requirements.txt
allure-behave==2.15.0
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.3.0
behave==1.3.3
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
coverage==7.10.6
cucumber-expressions==18.0.1
cucumber-tag-expressions==6.2.0
execnet==2.1.1
Faker==37.6.0
gherkin-official==29.0.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
parse_type==0.6.6
pluggy==1.6.0
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.2
pytest-bdd==8.1.0
pytest-cov==7.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-rerunfailures==16.0.1
pytest-xdist==3.8.0
python-dotenv==1.1.1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
